move `ConnectionContext` out of tsx file
