+++ utils/ataTools.tsx
@@ -5,7 +5,7 @@ import {
 } from '@solana/spl-token'
 import { WalletAdapter } from '@solana/wallet-adapter-base'
 import { Connection, PublicKey, Transaction } from '@solana/web3.js'
-import { ConnectionContext } from 'stores/useWalletStore'
+import type { ConnectionContext } from 'utils/connection'
 import { sendTransaction } from './send'
 import { tryGetAta, isExistingTokenAccount } from './validations'
